+++ pytest/test_genie_approx.py
@@ -16,11 +16,11 @@ rpy2.robjects.numpy2ri.activate()
 path = "benchmark_data"
 
 def test_genie(metric='euclidean'):
-    for dataset in ["s1", "Aggregation", "unbalance", "h2mg_64_50"]:#, "h2mg_1024_50", "t4_8k", "bigger"]:
+    for dataset in ["t4_8k", "h2mg_64_50", "h2mg_1024_50", "bigger"]:#["s1", "Aggregation", "unbalance", "h2mg_64_50"]:#, "h2mg_1024_50", "t4_8k", "bigger"]:
         if dataset == "bigger":
             np.random.seed(123)
-            n = 10_000
-            X = np.random.normal(size=(n,2))
+            n = 100_000
+            X = np.random.normal(size=(n,32))
             labels = np.random.choice(np.r_[1,2], n)
         else:
             X = np.loadtxt("%s/%s.data.gz" % (path,dataset), ndmin=2)
@@ -28,18 +28,17 @@ def test_genie(metric='euclidean'):
 
         label_counts = np.unique(labels,return_counts=True)[1]
         k = len(label_counts)
-
         # center X + scale (NOT: standardize!)
         X = (X-X.mean(axis=0))/X.std(axis=None, ddof=1)
         X += np.random.normal(0, 0.0001, X.shape)
 
-        for g in [0.01, 0.3, 0.5, 0.7, 1.0]:
+        for g in [0.01, 1.0]:
             gc.collect()
 
             #D = scipy.spatial.distance.pdist(X)
             #D = scipy.spatial.distance.squareform(D)
 
-            print("%-20s g=%.2f n=%5d d=%2d"%(dataset,g,X.shape[0],X.shape[1]), end="\t")
+            print("%-20s g=%.2f n=%7d d=%4d"%(dataset,g,X.shape[0],X.shape[1]), end="\t")
 
             t01 = time.time()
             res1 = Genie(k, g, exact=True, nn_params=dict(metric=metric)).fit_predict(X)+1
@@ -61,11 +60,14 @@ def test_genie(metric='euclidean'):
 
             print("t_rel=%.3f" % ((t11-t01)/(t12-t02),), end="\t")
 
-            # t0 = time.time()
-            # res3 = Genie(k, g, exact=False, n_neighbors=64, nn_params=dict(metric=metric, algorithm='brute')).fit_predict(X)+1
-            # print("t_py2=%.3f" % (time.time()-t0), end="\t")
-            # ari = adjusted_rand_score(res1, res3)
-            # print("ARI2=%.3f" % ari, end="\t")
+            t03 = time.time()
+            res3 = Genie(k, g, exact=False, n_neighbors=32, nn_params=dict(metric=metric)).fit_predict(X)+1
+            t13 = time.time()
+            print("t_py2=%.3f" % (t13-t03), end="\t")
+            print("t_rel=%.3f" % ((t03-t13)/(t01-t11),), end="\t")
+
+            ari = adjusted_rand_score(res1, res3)
+            print("ARI2=%.3f" % ari, end="\t")
 
             res1, res2 = None, None
             print("")
@@ -74,5 +76,5 @@ def test_genie(metric='euclidean'):
 if __name__ == "__main__":
     print("**Euclidean**")
     test_genie('euclidean')
-    print("**Manhattan**")
-    test_genie('manhattan')
+    #print("**Manhattan**")
+    #test_genie('manhattan')